[Design] asset 및 폰트, 윈도우 사이즈 수정
