+++ Cote/Cote/ContentView.swift
@@ -9,13 +9,8 @@ import SwiftUI
 
 struct ContentView: View {
     var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, cote")
-        }
-        .padding()
+        
+        Sidebar()
     }
 }
